plgx-esp/requirements/dev.txt
alembic==0.8.6
eventlet==0.24.1

amqp==1.4.9
anyjson==0.3.3
bcrypt==2.0.0
beautifulsoup4==4.4.1
billiard==3.3.0.23
blinker==1.4
celery==5.2.1
cffi==1.6.0
contextlib2==0.5.3
cssmin==0.2.0
enum34==1.1.6
factory-boy==2.7.0
fake-factory==0.5.7
Flask==0.10.1
Flask-Authorize==0.2.6
Flask-Assets==0.11
Flask-Bcrypt==0.7.1
Flask-Caching==1.7.1
Flask-Cors==3.0.8
Flask-DebugToolbar==0.10.0
flask-ldap3-login==0.9.9
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==1.8.0
flask-paginate==0.4.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-SSLify==0.1.5
Flask-WTF==0.12
funcsigs==1.0.2
ipaddress==1.0.16
itsdangerous==0.24
Jinja2==2.8
jsmin==2.2.1
kombu==3.0.35
ldap3==1.3.1
Mako==1.0.4
MarkupSafe==0.23
mock==2.0.0
oauthlib==1.1.1
OTXv2==1.3
pbr==1.9.1
psutil==5.8.0
#psycopg2==2.7.6.1
psycopg2-binary==2.8
py==1.4.31
pyasn1==0.1.9
pycparser==2.14
pytest==2.9.1
python-dateutil==2.5.3
python-editor==1.0
pytz==2016.4
raven==6.10.0
redis==2.10.5
requests==2.10.0
requests-oauthlib==0.6.1
scales==1.0.9
six==1.10.0
SQLAlchemy==1.0.12
waitress==0.9.0
unicodecsv==0.14.1
virustotal-api==1.1.10
webassets==0.11.1
WebOb==1.6.0
WebTest==2.0.21
Werkzeug==0.11.8
WTForms==2.1
pika==0.12.0

